fix: changes according to PR feedback
