+++ core/src/main/res/values/strings.xml
@@ -121,6 +121,7 @@
     <string name="core_accessibility_header_image_for">Header image for %1$s</string>
     <string name="core_accessibility_settings">Settings</string>
     <string name="core_accessibility_btn_back">Back</string>
+    <string name="core_accessibility_expandable_arrow">Expandable Arrow</string>
 
     <string name="core_video_download_to_device">Download to device</string>
     <string name="core_video_downloading_to_device">Downloading videos…</string>
