+++ profile/src/main/java/org/openedx/profile/presentation/anothersaccount/AnothersProfileFragment.kt
@@ -1,4 +1,4 @@
-package org.openedx.profile.presentation.anothers_account
+package org.openedx.profile.presentation.anothersaccount
 
 import android.content.res.Configuration.UI_MODE_NIGHT_NO
 import android.content.res.Configuration.UI_MODE_NIGHT_YES
