+++ profile/src/main/java/org/openedx/profile/presentation/anothersaccount/AnothersProfileViewModel.kt
@@ -1,4 +1,4 @@
-package org.openedx.profile.presentation.anothers_account
+package org.openedx.profile.presentation.anothersaccount
 
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
