+++ profile/src/main/java/org/openedx/profile/presentation/settings/SettingsFragment.kt
@@ -93,7 +93,6 @@ class SettingsFragment : Fragment() {
                     }
                 )
 
-
                 LaunchedEffect(logoutSuccess) {
                     if (logoutSuccess) {
                         viewModel.restartApp(requireActivity().supportFragmentManager)
